src/requirements/production.txt
django~=2.1
django-boogie
django-cors-headers
django-filter
django-picklefield
django-polymorphic
django-pyodbc
djangorestframework
django-rest-swagger
numpy~=1.15.1
pandas
pyodbc
psycopg2-binary
pymssql
scikit-learn
scipy
pycpfcnpj

